DaoFactory에 책임 전가를 하라
제어의 역전(`IoC`)로 제어 흐름을 반대로 하게 합니다.
UserDao의 역활을 DaoFactory이 대신 해줌으로써 UserDao는 수동적인 존재가 됩니다.
